requirements.txt
scikit-learn==1.6.1
jsonlines==4.0.0
json-repair
flake8==7.2.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-simplify==0.21.0
openai==1.59.9
pyyaml
oss2==2.19.1
torch==2.5.1
transformers==4.48.0
tqdm==4.67.1
scipy==1.13.1
numpy==2.0.2
pandas==2.2.3
faker
python-dateutil==2.9.0.post0
openpyxl==3.1.5